+++ pysui/abstracts/client_config.py
@@ -37,6 +37,11 @@ class ClientConfiguration(ABC):
     def active_address(self) -> AbstractType:
         """Return the active address from the client configuration."""
 
+    @classmethod
+    @abstractmethod
+    def generate_configuration(cls) -> "ClientConfiguration":
+        """Generate a default configuration."""
+
     @property
     def keystore_file(self) -> str:
         """Get the kestore filename."""