+++ samples/cmd_args.py
@@ -112,15 +112,14 @@ def _build_read_cmds(subparser) -> None:
     subp.set_defaults(subcommand="gas")
     # Object
     subp = subparser.add_parser("object", help="Show object by id")
-    subp.add_argument("--id", required=True, action=ValidateObjectID)
-    subp.add_argument("--json", required=False, help="Display output as json", action="store_true")
+    subp.add_argument("-i", "--id", required=True, action=ValidateObjectID)
+    subp.add_argument("-j", "--json", required=False, help="Display output as json", action="store_true")
     subp.set_defaults(subcommand="object")
     # Objects
     subp = subparser.add_parser("objects", help="Show all objects")
     subp.add_argument("-a", "--address", required=False, help="Objects for address", action=ValidateAddress)
     subp.add_argument("-j", "--json", required=False, help="Display output as json", action="store_true")
     obj_arg_group = subp.add_mutually_exclusive_group()
-    obj_arg_group.add_argument("-n", "--nft", help="Only show NFT objects", action="store_true")
     obj_arg_group.add_argument("-d", "--data", help="Only show data objects", action="store_true")
     subp.set_defaults(subcommand="objects")
     # RPC information